codeclimate duplication fix
- new modules created to import all navigation screens
